requirements.txt
# Shared config (editable install)
-e ../shared_config

# Core
numpy>=1.24.0,<2.0.0
polars>=0.20.0
pyyaml>=6.0

# PyTorch ecosystem
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Computer Vision
opencv-python>=4.9.0
Pillow>=10.0.0
albumentations>=1.3.0

# ML/DL
transformers>=4.36.0
scipy>=1.11.0
scikit-learn>=1.4.0

# Experiment Tracking
mlflow>=2.9.0
tensorboard>=2.14.0

# Training Utilities
pytorch-lightning>=2.1.0
tqdm>=4.65.0
hydra-core>=1.3.0

# Data & Viz
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
altair>=5.2.0

# Dev
ipykernel>=6.28.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0